+++ vue-frontend/src/components/ApplicationFrame/CommandBar/commandBarCore.ts
